Input event data should not be null when replace selection command is performed
As per spec
https://www.w3.org/TR/input-events-1/#overview, event data for input
events with certain inputType should be set with the data under
operation. Currently, the event data is null since the method
|TextDataForInputEvent| was not implemented in |ReplaceSelectionCommand|
class, which is implemented in this CL. |ReplaceSelectionCommand|
objects are populated with insertFromDrop, insertFromPaste,
insertReplacementText, insertTranspose and none inputType. This change
will set the data for input events with inputType insertFromDrop,
insertFromPaste and insertReplacementText leaving the insertTranspose
and none inputType data null as per the spec. For other input types,
other class objects derived from |EditCommand| are created, which is not
in the scope of this CL and can be handled separately.

Bug: 40737336
Change-Id: Ia7557f4d4fb239585be8fbcc5cdb489b77799c3f